refactor: admin handler
