+++ internal/api/admin.gotmpl
